Disable next.js scripts from being added to article pages
This fixes an Ezoic + Uber Grid integration problem.

We're using the new unstable_runtimeJS config parameter, which might
change in the future, so just keep an eye on this.

Also, replaced react-native <Image> tags with <img> tags because
the former would not render properly when next-js scripts were
excluded from the article pages.